requirements.txt
scikit-learn==1.5.0
google-genai==0.3.0
pypdf==4.2.0